pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "py-vuln-scout"
version = "0.1.0"
description = "A modular SAST tool for Python with regex and LLM-based vulnerability detection"
authors = [{name = "py-vuln-scout Contributors"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "typer[all]>=0.9.0",
    "pydantic>=2.5.0",
    "requests>=2.31.0",
    "libcst>=1.1.0",
    "jinja2>=3.1.0",
    "jsonschema>=4.20.0",
    "toml>=0.10.2",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "mypy>=1.7.0",
    "types-requests>=2.31.0",
    "types-PyYAML>=6.0.0",
    "types-toml>=0.10.0",
    "pre-commit>=3.5.0",
]

[project.scripts]
pvs = "py_vuln_scout.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
py_vuln_scout = [
    "rules/**/*.json",
    "rules/**/*.yaml",
    "rules/schema/*.json",
    "output/schemas/*.json",
]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_optional = true
strict_equality = true

[[tool.mypy.overrides]]
module = "libcst.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=py_vuln_scout --cov-report=term-missing --cov-report=html"
